Enum stc_ts_errors::ErrorKind

pub enum ErrorKind {

    NonObjectInSpread {
        span: Span,
        ty: Box<Type>,
    },
    NotExtendableType {
        span: Span,
    },
    InvalidInterfaceInheritance {
        span: Span,
        cause: Box<Error>,
    },
    TupleTooShort {
        span: Span,
    },
    VarDeclNotCompatible {
        span: Span,
        cause: Box<Error>,
    },
    IntrinsicIsBuiltinOnly {
        span: Span,
    },
    TypeParamsProvidedButCalleeIsNotGeneric {
        span: Span,
    },
    OptionalAndNonOptionalMethodPropertyMixed {
        span: Span,
    },
    UpdateArgMustBeVariableOrPropertyAccess {
        span: Span,
    },
    InterfaceNotCompatible {
        span: Span,
    },
    CannotUseTypeAsIndexIndex {
        span: Span,
    },
    CannotFunctionReturningNever {
        span: Span,
    },
    OnlyOneEnumCanOmitInit {
        span: Span,
    },
    ConstEnumMemberHasInfinityAsInit {
        span: Span,
    },
    ConstEnumMemberHasNaNAsInit {
        span: Span,
    },
    UnreachableCode {
        span: Span,
    },
    VarMayNotBeInitialized {
        span: Span,
    },
    ClassPropNotInitialized {
        span: Span,
    },
    DefinedWithAccessorInSuper {
        span: Span,
    },
    OptionInvalidForEs3 {
        span: Span,
    },
    ThisUsedBeforeCallingSuper {
        span: Span,
    },
    SuperUsedBeforeCallingSuper {
        span: Span,
    },
    SuperInNestedFunction {
        span: Span,
    },
    SuperNotCalled {
        span: Span,
    },
    CannotAccessAbstractMember {
        span: Span,
    },
    OperatorCannotBeAppliedToTypes {
        span: Span,
    },
    AbstractClassMethodShouldBeSequential {
        span: Span,
    },
    ClassMemberNotCompatibleWithStringIndexSignature {
        span: Span,
    },
    ClassMemberNotCompatibleWithNumericIndexSignature {
        span: Span,
    },
    AssignFailedBecauseTupleLengthDiffers {
        span: Span,
    },
    InvalidUsageOfNewTarget {
        span: Span,
    },
    ReturnPropertyOfIteratorMustBeMethod {
        span: Span,
    },
    NextOfIteratorShouldReturnTypeWithPropertyValue {
        span: Span,
    },
    CannotAssignToNamespace {
        span: Span,
    },
    RestArgMustBeVarOrMemberAccess {
        span: Span,
    },
    InvalidOperandOfIncDecOptionalProp {
        span: Span,
    },
    InvalidRestPatternInOptionalChain {
        span: Span,
    },
    InvalidLhsOfAssignOptionalProp {
        span: Span,
    },
    InvalidRestPatternInForIn {
        span: Span,
    },
    InvalidRestPatternInForOf {
        span: Span,
    },
    BindingPatNotAllowedInRestPatArg {
        span: Span,
    },
    DeleteOperandMustBeOptional {
        span: Span,
    },
    SuperCannotUseTypeArgs {
        span: Span,
    },
    TargetLacksConstructSignature {
        span: Span,
    },
    BlockScopedVarUsedBeforeInit {
        span: Span,
    },
    DuplicateDefaultExport {
        span: Span,
    },
    DuplicateExport {
        span: Span,
    },
    DuplicateFnImpl {
        span: Span,
    },
    DuplicateConstructor {
        span: Span,
    },
    ModuleNotFound {
        span: Span,
    },
    TooManyAsterisk {
        span: Span,
    },
    DuplicateVar {
        name: Id,
        span: Span,
    },
    ClassNameCannotBeObjectWhenTargetingEs5WithModule {
        span: Span,
    },
    NotArrayType {
        span: Span,
    },
    NotArrayTypeNorStringType {
        span: Span,
    },
    NotArrayTypeNorStringTypeButDownlevelIterationWouldWork {
        span: Span,
    },
    ForOfStringUsedInEs3 {
        span: Span,
    },
    LetOrConstIsNotValidIdInLetOrConstVarDecls {
        span: Span,
    },
    InvalidExprOfLhsOfForIn {
        span: Span,
    },
    InvalidExprOfLhsOfForOf {
        span: Span,
    },
    WrongTypeForLhsOfForInLoop {
        span: Span,
    },
    DestructuringBindingNotAllowedInLhsOfForIn {
        span: Span,
    },
    TypeAnnOnLhsOfForInLoops {
        span: Span,
    },
    TypeAnnOnLhsOfForOfLoops {
        span: Span,
    },
    CannotAccessPrivatePropertyFromOutside {
        span: Span,
    },
    OptionalChainCannotContainPrivateIdentifier {
        span: Span,
    },
    CannotDeletePrivateProperty {
        span: Span,
    },
    ConstructorIsKeyword {
        span: Span,
    },
    PrivateIdUsedAsMethodName {
        span: Span,
    },
    UndefinedOrNullIsNotValidOperand {
        span: Span,
    },
    ThisInConstructorParam {
        span: Span,
    },
    ThisInStaticPropertyInitializer {
        span: Span,
    },
    InvalidSuperClass {
        span: Span,
    },
    WithStmtNotSupported {
        span: Span,
    },
    NamespaceNotFound {
        name: Box<Name>,
        ctxt: ModuleId,
        type_args: Option<Box<TypeParamInstantiation>>,
        span: Span,
    },
    EnumMemberIdCannotBeNumber {
        span: Span,
    },
    InvalidLhsOfAssign {
        span: Span,
    },
    ImplicitReturnType {
        span: Span,
    },
    IncompatibleFnOverload {
        span: Span,
        cause: Box<Error>,
    },
    InitializerDisallowedInAmbientContext {
        span: Span,
    },
    InvalidClassName {
        span: Span,
    },
    NoSuchVarForShorthand {
        span: Span,
        name: Id,
    },
    NoMatchingOverload {
        span: Span,
    },
    InvalidInterfaceName {
        span: Span,
    },
    CannotCallWithNewNonVoidFunction {
        span: Span,
    },
    DuplicateProperty {
        span: Span,
    },
    CannotExportNonLocalVar {
        span: Span,
    },
    StaticPropertyCannotBeNamedPrototype {
        span: Span,
    },
    SelfReferentialSuperClass {
        span: Span,
    },
    NotConstructorType {
        span: Span,
    },
    ExportMixedWithLocal {
        span: Span,
    },
    MixedDefaultExports {
        span: Span,
    },
    ClassIncorrectlyImplementsInterface {
        span: Span,
    },
    StackOverflow {
        span: Span,
    },
    InvalidImplOfInterface {
        span: Span,
        cause: Box<Error>,
    },
    StaticMemberCannotUseTypeParamOfClass {
        span: Span,
    },
    DeclaringTypeParamReferencedByComputedPropName {
        span: Span,
    },
    CannotReferenceThisInComputedPropName {
        span: Span,
    },
    CannotReferenceSuperInComputedPropName {
        span: Span,
    },
    ThisRefToModuleOrNamespace {
        span: Span,
    },
    SuperInClassWithoutSuper {
        span: Span,
    },
    SuperCanBeOnlyReferencedInDerivedClass {
        span: Span,
    },
    GeneratorCannotHaveVoidAsReturnType {
        span: Span,
    },
    NoSuchVarButThisHasSuchProperty {
        span: Span,
        name: Id,
    },
    DestructuringAssignInAmbientContext {
        span: Span,
    },
    OptionalBindingPatternInImplSignature {
        span: Span,
    },
    NullishCoalescingMixedWithLogicalWithoutParen {
        span: Span,
    },
    SwitchCaseTestNotCompatible {
        span: Span,
        disc: Box<Type>,
        test: Box<Type>,
    },
    EnumCannotBeLValue {
        span: Span,
    },
    ExprInvalidForUpdateArg {
        span: Span,
    },
    TypeInvalidForUpdateArg {
        span: Span,
        ty: Box<Type>,
    },
    PrivatePropertyIsDifferent {
        span: Span,
    },
    PrivateMethodIsDifferent {
        span: Span,
    },
    CannotCompareWithOp {
        span: Span,
        op: BinaryOp,
        left: Box<Type>,
        right: Box<Type>,
    },
    InvalidBinaryOp {
        span: Span,
        op: BinaryOp,
        left: Box<Type>,
        right: Box<Type>,
    },
    NoSuchEnumVariant {
        span: Span,
        name: JsWord,
    },
    ObjectIsPossiblyNull {
        span: Span,
    },
    ObjectIsPossiblyUndefined {
        span: Span,
    },
    ObjectIsPossiblyUndefinedWithType {
        span: Span,
        ty: Box<Type>,
    },
    ObjectIsPossiblyNullOrUndefined {
        span: Span,
    },
    CannotAssignAbstractConstructorToNonAbstractConstructor {
        span: Span,
    },
    InvalidUseOfConstEnum {
        span: Span,
    },
    ComputedMemberInEnumWithStrMember {
        span: Span,
    },
    CannotCreateInstanceOfAbstractClass {
        span: Span,
    },
    NotSatisfyConstraint {
        span: Span,
        left: Box<Type>,
        right: Box<Type>,
    },
    WrongArgType {
        span: Span,
        inner: Box<Error>,
    },
    ImportFailed {
        span: Span,
        orig: Id,
        id: Id,
    },
    NonStringDynamicImport {
        span: Span,
    },
    ExportFailed {
        span: Span,
        orig: Id,
        id: Id,
    },
    ExportAllFailed {
        span: Span,
    },
    NoSuchPropertyInThis {
        span: Span,
    },
    NoSuchPropertyInClass {
        span: Span,
        class_name: Option<Id>,
        prop: Key,
    },
    TypeParameterCountMismatch {
        span: Span,
        min: usize,
        max: usize,
        actual: usize,
    },
    ParameterCountMismatch {
        span: Span,
        min: usize,
        max: usize,
        actual: usize,
    },
    NoSuchPropertyInModule {
        span: Span,
        name: Box<Key>,
    },
    ReturnRequired {
        span: Span,
    },
    ConstructorRequired {
        span: Span,
        lhs: Span,
        rhs: Span,
    },
    CannotAssignToNonVariable {
        span: Span,
        ty: Box<Type>,
    },
    CannotAssignToModule {
        span: Span,
    },
    CannotAssignToClass {
        span: Span,
    },
    CannotAssignToEnum {
        span: Span,
    },
    CannotAssignToFunction {
        span: Span,
    },
    AssignedWrapperToPrimitive {
        span: Span,
    },
    AccessibilityDiffers {
        span: Span,
    },
    InvalidInitInConstEnum {
        span: Span,
    },
    InvalidTupleCast {
        span: Span,
        left: Span,
        right: Span,
    },
    NoOverlap {
        span: Span,
        value: bool,
        left: Box<Type>,
        right: Box<Type>,
    },
    CannotAssignToReadonlyProperty {
        span: Span,
    },
    ReadOnly {
        span: Span,
    },
    ImplicitAny {
        span: Span,
    },
    ImplicitAnyBecauseNoIndexSignatureExists {
        span: Span,
    },
    ImplicitAnyBecauseIndexTypeIsWrong {
        span: Span,
    },
    ImplicitAnyBecauseOfSelfRef {
        span: Span,
    },
    TupleAssignError {
        span: Span,
        errors: Vec<Error>,
    },
    Errors {
        span: Span,
        errors: Vec<Error>,
    },
    RedeclaredVarWithDifferentType {
        span: Span,
    },
    NoSuchType {
        span: Span,
        name: Id,
    },
    NoSuchTypeButVarExists {
        span: Span,
        name: Id,
    },
    InvalidUseOfArgumentsInEs3OrEs5 {
        span: Span,
    },
    ArgumentsCannotBeUsedInAsyncFnInEs3OrEs5 {
        span: Span,
    },
    NoSuchVar {
        span: Span,
        name: Id,
    },
    CannotExtendTypeOnlyItem {
        span: Span,
        name: Id,
    },
    TypeUsedAsVar {
        span: Span,
        name: Id,
    },
    DuplicateName {
        name: Id,
        span: Span,
    },
    DuplicateNameWithoutName {
        span: Span,
    },
    UselessSeqExpr {
        span: Span,
    },
    ClassPropertyInitRequired {
        span: Span,
    },
    ReferencedInInit {
        span: Span,
    },
    NotGeneric {
        span: Span,
    },
    Unknown {
        span: Span,
    },
    NoSuchPropertyWhileDeclWithBindingPat {
        span: Span,
    },
    NoSuchProperty {
        span: Span,
        obj: Option<Box<Type>>,
        prop: Option<Box<Key>>,
    },
    NoInitAndNoDefault {
        span: Span,
    },
    TooManyTupleElements {
        span: Span,
    },
    NotTuple {
        span: Span,
    },
    NotVariable {
        span: Span,
        left: Span,
        ty: Option<Box<Type>>,
    },
    TypeNotFound {
        name: Box<Name>,
        ctxt: ModuleId,
        type_args: Option<Box<TypeParamInstantiation>>,
        span: Span,
    },
    TS2378 {
        span: Span,
    },
    ConstEnumNonIndexAccess {
        span: Span,
    },
    TupleIndexError {
        span: Span,
        len: u64,
        index: i64,
    },
    NegativeTupleIndex {
        span: Span,
        len: u64,
        index: i64,
    },
    InvalidLValue {
        span: Span,
    },
    Unimplemented {
        span: Span,
        msg: String,
    },
    ResolvedFailed {
        span: Span,
        base: Box<PathBuf>,
        src: JsWord,
    },
    MissingFields {
        span: Span,
        fields: Vec<TypeElement>,
    },
    AssignFailed {
        span: Span,
        left: Box<Type>,
        right_ident: Option<Span>,
        right: Box<Type>,
        cause: Vec<Error>,
    },
    AssignFailedDueToAccessibility {
        span: Span,
    },
    AssignFailedDueToOptionalityDifference {
        span: Span,
    },
    ObjectAssignFailed {
        span: Span,
        errors: Vec<Error>,
    },
    SimpleAssignFailed {
        span: Span,
        cause: Option<Box<Error>>,
    },
    SimpleAssignFailedWithCause {
        span: Span,
        cause: Vec<Error>,
    },
    InvalidAssignmentOfArray {
        span: Span,
    },
    UnionError {
        span: Span,
        errors: Vec<Error>,
    },
    IntersectionError {
        span: Span,
        error: Box<Error>,
    },
    CannotAssignToThis {
        span: Span,
    },
    MayBeUndefined {
        span: Span,
    },
    UndefinedSymbol {
        sym: Id,
        span: Span,
    },
    ModuleLoadFailed {
        span: Span,
        errors: Errors,
    },
    NoSuchExport {
        span: Span,
        items: Vec<Id>,
    },
    NoNewSignature {
        span: Span,
        callee: Box<Type>,
    },
    NoConstructablePropertyWithName {
        span: Span,
        obj: Box<Type>,
        key: Box<Key>,
    },
    NoCallSignature {
        span: Span,
        callee: Box<Type>,
    },
    WrongTypeParams {
        span: Span,
        callee: Span,
        expected: RangeInclusive<usize>,
        actual: usize,
    },
    WrongParams {
        span: Span,
        callee: Span,
        expected: RangeInclusive<usize>,
        actual: usize,
    },
    InvalidEnumInit {
        span: Span,
    },
    TS1016 {
        span: Span,
    },
    TS1063 {
        span: Span,
    },
    TS1094 {
        span: Span,
    },
    TS1095 {
        span: Span,
    },
    TS1168 {
        span: Span,
    },
    TS1169 {
        span: Span,
    },
    TS1183 {
        span: Span,
    },
    TS1318 {
        span: Span,
    },
    TS1319 {
        span: Span,
    },
    ExportEqualsMixedWithOtherExports {
        span: Span,
    },
    AnyTypeUsedAsCalleeWithTypeArgs {
        span: Span,
    },
    InvalidLhsOfInOperator {
        span: Span,
    },
    InvalidRhsForInOperator {
        span: Span,
        ty: Box<Type>,
    },
    WrongTypeForLhsOfNumericOperation {
        span: Span,
    },
    WrongTypeForRhsOfNumericOperation {
        span: Span,
        ty: Box<Type>,
    },
    TS2365 {
        span: Span,
    },
    TS2370 {
        span: Span,
    },
    WrongOverloadSignature {
        span: Span,
    },
    TS1166 {
        span: Span,
    },
    TS1345 {
        span: Span,
    },
    TS2353 {
        span: Span,
    },
    ConstructorImplMissingOrNotFollowedByDecl {
        span: Span,
    },
    FnImplMissingOrNotFollowedByDecl {
        span: Span,
    },
    InvalidTypeForComputedProperty {
        span: Span,
        ty: Box<Type>,
    },
    ParamPropIsNotAllowedInAmbientConstructor {
        span: Span,
    },
    TS2389 {
        span: Span,
    },
    TS2447 {
        span: Span,
    },
    ClassDoesNotImplementMember {
        span: Span,
        key: Box<Key>,
    },
    TS2531 {
        span: Span,
    },
    TS2567 {
        span: Span,
    },
    TS2585 {
        span: Span,
    },
    TS2704 {
        span: Span,
    },
    InvalidLhsInInstanceOf {
        span: Span,
        ty: Box<Type>,
    },
    InvalidRhsInInstanceOf {
        span: Span,
        ty: Box<Type>,
    },
    NumericOpToSymbol {
        span: Span,
    },
    InvalidNumericOperand {
        span: Span,
    },
    UpdateOpToSymbol {
        span: Span,
        op: UpdateOp,
    },
    UnknownPropertyInObjectLiteralAssignment {
        span: Span,
    },
    NonOverlappingTypeCast {
        span: Span,
    },
    InvalidOperatorForLhs {
        span: Span,
        op: AssignOp,
    },
    InvalidOpAssign {
        span: Span,
        op: AssignOp,
        lhs: Box<Type>,
        rhs: Box<Type>,
    },
    AssignOpCannotBeApplied {
        span: Span,
        op: AssignOp,
    },
    ExpectedNArgsButGotM {
        span: Span,
        min: usize,
        max: Option<usize>,
    },
    ExpectedAtLeastNArgsButGotM {
        span: Span,
        min: usize,
    },
    ExpectedAtLeastNArgsButGotMOrMore {
        span: Span,
        min: usize,
    },
    ExpectedNArgsButGotMOrMore {
        span: Span,
    },
    SpreadMustBeTupleOrPassedToRest {
        span: Span,
    },
    InvalidDeleteOperand {
        span: Span,
    },
    NoMethodNamedNext {
        span: Span,
    },
    NoCallablePropertyWithName {
        span: Span,
        obj: Box<Type>,
        key: Box<Key>,
    },
    MustBeArray {
        span: Span,
    },
    MustHaveSymbolIteratorThatReturnsIteratorOrMustBeArray {
        span: Span,
    },
    MustHaveSymbolIteratorThatReturnsIterator {
        span: Span,
    },
    RightHandSideMustBeObject {
        span: Span,
        ty: Box<Type>,
    },
    MustHaveSymbolAsyncIteratorThatReturnsIterator {
        span: Span,
    },
    NoSuchConstructor {
        span: Span,
        key: Box<Key>,
    },
    AbstractAndConcreteIsMixed {
        span: Span,
    },
    SuperCanOnlyAccessPublicAndProtectedMethod {
        span: Span,
    },
    ShouldBeStaticMethod {
        span: Span,
    },
    ShouldBeInstanceMethod {
        span: Span,
    },
    RestPropertyNotLast {
        span: Span,
    },
    ClassConstructorPrivate {
        span: Span,
    },
    ClassConstructorProtected {
        span: Span,
    },
    InvalidExtendDueToConstructorPrivate {
        span: Span,
    },
    DuplicatePrivateStaticInstance {
        span: Span,
    },
    ExportAmbientModule {
        span: Span,
    },
    IsTypeUnknown {
        span: Span,
    },
    RestTypeNotFromObject {
        span: Span,
    },
    UsePropBeforeInit {
        span: Span,
        obj: Option<Box<Type>>,
        prop: Option<Box<Key>>,
    },

}

Variants

NonObjectInSpread

Fields

span: Span

ty: Box<Type>

TS2698

NotExtendableType

Fields

span: Span

TS2312

InvalidInterfaceInheritance

Fields

span: Span

cause: Box<Error>

TS2430

TupleTooShort

Fields

span: Span

TS2339

VarDeclNotCompatible

Fields

span: Span

cause: Box<Error>

TS2403

IntrinsicIsBuiltinOnly

Fields

span: Span

TS2795

TypeParamsProvidedButCalleeIsNotGeneric

Fields

span: Span

TS2347

OptionalAndNonOptionalMethodPropertyMixed

Fields

span: Span

TS2386

UpdateArgMustBeVariableOrPropertyAccess

Fields

span: Span

TS2357

InterfaceNotCompatible

Fields

span: Span

TS2320

CannotUseTypeAsIndexIndex

Fields

span: Span

TS2538

CannotFunctionReturningNever

Fields

span: Span

TS2534 A function returning ‘never’ cannot have a reachable end point.

OnlyOneEnumCanOmitInit

Fields

span: Span

TS2432

ConstEnumMemberHasInfinityAsInit

Fields

span: Span

TS2477

ConstEnumMemberHasNaNAsInit

Fields

span: Span

TS2478

UnreachableCode

Fields

span: Span

TS7027

VarMayNotBeInitialized

Fields

span: Span

TS2454

ClassPropNotInitialized

Fields

span: Span

TS2564

DefinedWithAccessorInSuper

Fields

span: Span

TS2610

OptionInvalidForEs3

Fields

span: Span

TS5048

ThisUsedBeforeCallingSuper

Fields

span: Span

TS17009

SuperUsedBeforeCallingSuper

Fields

span: Span

TS17011

SuperInNestedFunction

Fields

span: Span

TS2337

SuperNotCalled

Fields

span: Span

TS2377

CannotAccessAbstractMember

Fields

span: Span

TS2513

OperatorCannotBeAppliedToTypes

Fields

span: Span

TS2365

AbstractClassMethodShouldBeSequential

Fields

span: Span

TS2516

ClassMemberNotCompatibleWithStringIndexSignature

Fields

span: Span

TS2411

ClassMemberNotCompatibleWithNumericIndexSignature

Fields

span: Span

TS2411

AssignFailedBecauseTupleLengthDiffers

Fields

span: Span

TS2322

InvalidUsageOfNewTarget

Fields

span: Span

TS17013

ReturnPropertyOfIteratorMustBeMethod

Fields

span: Span

TS2767

NextOfIteratorShouldReturnTypeWithPropertyValue

Fields

span: Span

TS2490

CannotAssignToNamespace

Fields

span: Span

TS2631

RestArgMustBeVarOrMemberAccess

Fields

span: Span

TS2701

InvalidOperandOfIncDecOptionalProp

Fields

span: Span

TS2777

InvalidRestPatternInOptionalChain

Fields

span: Span

TS2778

InvalidLhsOfAssignOptionalProp

Fields

span: Span

TS2779

InvalidRestPatternInForIn

Fields

span: Span

TS2780

InvalidRestPatternInForOf

Fields

span: Span

TS2781

BindingPatNotAllowedInRestPatArg

Fields

span: Span

TS2501

DeleteOperandMustBeOptional

Fields

span: Span

TS2790

SuperCannotUseTypeArgs

Fields

span: Span

TS2754

TargetLacksConstructSignature

Fields

span: Span

TS7009

BlockScopedVarUsedBeforeInit

Fields

span: Span

TS2448

DuplicateDefaultExport

Fields

span: Span

TS2528

DuplicateExport

Fields

span: Span

TS2323

DuplicateFnImpl

Fields

span: Span

TS2393

DuplicateConstructor

Fields

span: Span

TS2392

ModuleNotFound

Fields

span: Span

TS2307

TooManyAsterisk

Fields

span: Span

TS5061

DuplicateVar

Fields

name: Id

span: Span

TS2451

ClassNameCannotBeObjectWhenTargetingEs5WithModule

Fields

span: Span

TS2725

NotArrayType

Fields

span: Span

TS2461

NotArrayTypeNorStringType

Fields

span: Span

TS2495

NotArrayTypeNorStringTypeButDownlevelIterationWouldWork

Fields

span: Span

TS2569

ForOfStringUsedInEs3

Fields

span: Span

TS2494

LetOrConstIsNotValidIdInLetOrConstVarDecls

Fields

span: Span

TS2480

InvalidExprOfLhsOfForIn

Fields

span: Span

TS2406

InvalidExprOfLhsOfForOf

Fields

span: Span

TS2487

WrongTypeForLhsOfForInLoop

Fields

span: Span

TS2405

DestructuringBindingNotAllowedInLhsOfForIn

Fields

span: Span

TS2491

TypeAnnOnLhsOfForInLoops

Fields

span: Span

TS2404

TypeAnnOnLhsOfForOfLoops

Fields

span: Span

TS2483

CannotAccessPrivatePropertyFromOutside

Fields

span: Span

TS18013

OptionalChainCannotContainPrivateIdentifier

Fields

span: Span

TS18030

CannotDeletePrivateProperty

Fields

span: Span

TS18011

ConstructorIsKeyword

Fields

span: Span

TS18012

PrivateIdUsedAsMethodName

Fields

span: Span

TS18022

UndefinedOrNullIsNotValidOperand

Fields

span: Span

TS18050

ThisInConstructorParam

Fields

span: Span

TS2333

ThisInStaticPropertyInitializer

Fields

span: Span

TS2334

InvalidSuperClass

Fields

span: Span

TS2507

WithStmtNotSupported

Fields

span: Span

TS2410

NamespaceNotFound

Fields

name: Box<Name>

ctxt: ModuleId

type_args: Option<Box<TypeParamInstantiation>>

span: Span

TS2503

EnumMemberIdCannotBeNumber

Fields

span: Span

TS2452

InvalidLhsOfAssign

Fields

span: Span

TS2364

ImplicitReturnType

Fields

span: Span

TS7010

IncompatibleFnOverload

Fields

span: Span

cause: Box<Error>

TS2394

InitializerDisallowedInAmbientContext

Fields

span: Span

TS2371

InvalidClassName

Fields

span: Span

TS2414

NoSuchVarForShorthand

Fields

span: Span

name: Id

TS18004

NoMatchingOverload

Fields

span: Span

TS2769

InvalidInterfaceName

Fields

span: Span

TS2427

CannotCallWithNewNonVoidFunction

Fields

span: Span

DuplicateProperty

Fields

span: Span

TS2300

CannotExportNonLocalVar

Fields

span: Span

TS2661

StaticPropertyCannotBeNamedPrototype

Fields

span: Span

TS2699

SelfReferentialSuperClass

Fields

span: Span

TS2506

NotConstructorType

Fields

span: Span

TS2507

ExportMixedWithLocal

Fields

span: Span

TS2395

MixedDefaultExports

Fields

span: Span

TS2652

ClassIncorrectlyImplementsInterface

Fields

span: Span

TS2420

StackOverflow

Fields

span: Span

InvalidImplOfInterface

Fields

span: Span

cause: Box<Error>

TS2420

StaticMemberCannotUseTypeParamOfClass

Fields

span: Span

TS2302

DeclaringTypeParamReferencedByComputedPropName

Fields

span: Span

TS2467

CannotReferenceThisInComputedPropName

Fields

span: Span

TS2465

CannotReferenceSuperInComputedPropName

Fields

span: Span

TS2466

ThisRefToModuleOrNamespace

Fields

span: Span

TS2331

SuperInClassWithoutSuper

Fields

span: Span

TS2335

SuperCanBeOnlyReferencedInDerivedClass

Fields

span: Span

TS2660

GeneratorCannotHaveVoidAsReturnType

Fields

span: Span

TS2505

NoSuchVarButThisHasSuchProperty

Fields

span: Span

name: Id

TS2663

DestructuringAssignInAmbientContext

Fields

span: Span

OptionalBindingPatternInImplSignature

Fields

span: Span

TS2463

NullishCoalescingMixedWithLogicalWithoutParen

Fields

span: Span

TS5076

SwitchCaseTestNotCompatible

Fields

span: Span

disc: Box<Type>

test: Box<Type>

TS2678

EnumCannotBeLValue

Fields

span: Span

TS2540

ExprInvalidForUpdateArg

Fields

span: Span

TS2357

TypeInvalidForUpdateArg

Fields

span: Span

ty: Box<Type>

Type of the arguments.

TS2356

PrivatePropertyIsDifferent

Fields

span: Span

PrivateMethodIsDifferent

Fields

span: Span

CannotCompareWithOp

Fields

span: Span

op: BinaryOp

left: Box<Type>

right: Box<Type>

InvalidBinaryOp

Fields

span: Span

op: BinaryOp

left: Box<Type>

right: Box<Type>

NoSuchEnumVariant

Fields

span: Span

name: JsWord

TS2339

ObjectIsPossiblyNull

Fields

span: Span

ObjectIsPossiblyUndefined

Fields

span: Span

ObjectIsPossiblyUndefinedWithType

Fields

span: Span

ty: Box<Type>

ObjectIsPossiblyNullOrUndefined

Fields

span: Span

CannotAssignAbstractConstructorToNonAbstractConstructor

Fields

span: Span

InvalidUseOfConstEnum

Fields

span: Span

ComputedMemberInEnumWithStrMember

Fields

span: Span

CannotCreateInstanceOfAbstractClass

Fields

span: Span

NotSatisfyConstraint

Fields

span: Span

left: Box<Type>

right: Box<Type>

TS2344

WrongArgType

Fields

span: Span

Span of argument.

inner: Box<Error>

TS2345

ImportFailed

Fields

span: Span

orig: Id

id: Id

NonStringDynamicImport

Fields

span: Span

TS7036

ExportFailed

Fields

span: Span

orig: Id

id: Id

ExportAllFailed

Fields

span: Span

NoSuchPropertyInThis

Fields

span: Span

NoSuchPropertyInClass

Fields

span: Span

class_name: Option<Id>

prop: Key

TypeParameterCountMismatch

Fields

span: Span

min: usize

max: usize

actual: usize

ParameterCountMismatch

Fields

span: Span

min: usize

max: usize

actual: usize

NoSuchPropertyInModule

Fields

span: Span

name: Box<Key>

ReturnRequired

Fields

span: Span

Span of the return type.

TS2355

ConstructorRequired

Fields

span: Span

lhs: Span

rhs: Span

CannotAssignToNonVariable

Fields

span: Span

ty: Box<Type>

TS2539

CannotAssignToModule

Fields

span: Span

TS2708

CannotAssignToClass

Fields

span: Span

TS2629

CannotAssignToEnum

Fields

span: Span

TS2628

CannotAssignToFunction

Fields

span: Span

TS2630

AssignedWrapperToPrimitive

Fields

span: Span

TS2322

AccessibilityDiffers

Fields

span: Span

TS2322

InvalidInitInConstEnum

Fields

span: Span

TS2474

InvalidTupleCast

Fields

span: Span

left: Span

right: Span

TS2352

NoOverlap

Fields

span: Span

value: bool

left: Box<Type>

right: Box<Type>

TS2367

CannotAssignToReadonlyProperty

Fields

span: Span

ReadOnly

Fields

span: Span

ImplicitAny

Fields

span: Span

ImplicitAnyBecauseNoIndexSignatureExists

Fields

span: Span

TS7052

ImplicitAnyBecauseIndexTypeIsWrong

Fields

span: Span

TS7053

ImplicitAnyBecauseOfSelfRef

Fields

span: Span

TS7022

TupleAssignError

Fields

span: Span

errors: Vec<Error>

Errors

Fields

span: Span

errors: Vec<Error>

RedeclaredVarWithDifferentType

Fields

span: Span

NoSuchType

Fields

span: Span

name: Id

TS2304

NoSuchTypeButVarExists

Fields

span: Span

name: Id

TS2749

InvalidUseOfArgumentsInEs3OrEs5

Fields

span: Span

TS2496

ArgumentsCannotBeUsedInAsyncFnInEs3OrEs5

Fields

span: Span

TS2522

NoSuchVar

Fields

span: Span

name: Id

CannotExtendTypeOnlyItem

Fields

span: Span

name: Id

TS2689

TypeUsedAsVar

Fields

span: Span

name: Id

TS2693

DuplicateName

Fields

name: Id

span: Span

TS2300

DuplicateNameWithoutName

Fields

span: Span

TS2300

UselessSeqExpr

Fields

span: Span

TS2695

ClassPropertyInitRequired

Fields

span: Span

ReferencedInInit

Fields

span: Span

NotGeneric

Fields

span: Span

Unknown

Fields

span: Span

NoSuchPropertyWhileDeclWithBindingPat

Fields

span: Span

NoSuchProperty

Fields

span: Span

obj: Option<Box<Type>>

prop: Option<Box<Key>>

NoInitAndNoDefault

Fields

span: Span

TooManyTupleElements

Fields

span: Span

NotTuple

Fields

span: Span

NotVariable

Fields

span: Span

left: Span

ty: Option<Box<Type>>

TS2539

TypeNotFound

Fields

name: Box<Name>

ctxt: ModuleId

type_args: Option<Box<TypeParamInstantiation>>

span: Span

TS2304

TS2378

Fields

span: Span

TS2378

ConstEnumNonIndexAccess

Fields

span: Span

TS2476

TupleIndexError

Fields

span: Span

len: u64

index: i64

TS2493

NegativeTupleIndex

Fields

span: Span

len: u64

index: i64

TS2514

InvalidLValue

Fields

span: Span

Unimplemented

Fields

span: Span

msg: String

ResolvedFailed

Fields

span: Span

base: Box<PathBuf>

src: JsWord

MissingFields

Fields

span: Span

fields: Vec<TypeElement>

AssignFailed

Fields

span: Span

left: Box<Type>

right_ident: Option<Span>

right: Box<Type>

cause: Vec<Error>

TS2322

AssignFailedDueToAccessibility

Fields

span: Span

TS2322

AssignFailedDueToOptionalityDifference

Fields

span: Span

TS2322

ObjectAssignFailed

Fields

span: Span

errors: Vec<Error>

SimpleAssignFailed

Fields

span: Span

cause: Option<Box<Error>>

SimpleAssignFailedWithCause

Fields

span: Span

cause: Vec<Error>

InvalidAssignmentOfArray

Fields

span: Span

UnionError

Fields

span: Span

errors: Vec<Error>

a or b or c

IntersectionError

Fields

span: Span

error: Box<Error>

CannotAssignToThis

Fields

span: Span

MayBeUndefined

Fields

span: Span

Span of the variable

UndefinedSymbol

Fields

sym: Id

span: Span

ModuleLoadFailed

Fields

span: Span

Span of the import statement.

errors: Errors

NoSuchExport

Fields

span: Span

items: Vec<Id>

NoNewSignature

Fields

span: Span

callee: Box<Type>

TS2351

NoConstructablePropertyWithName

Fields

span: Span

obj: Box<Type>

key: Box<Key>

TS2348

NoCallSignature

Fields

span: Span

callee: Box<Type>

TS2349

WrongTypeParams

Fields

span: Span

Span of caller.

callee: Span

Span of callee.

expected: RangeInclusive<usize>

actual: usize

WrongParams

Fields

span: Span

Span of caller.

callee: Span

Span of callee.

expected: RangeInclusive<usize>

actual: usize

InvalidEnumInit

Fields

span: Span

TS1016

Fields

span: Span

TS1063

Fields

span: Span

TS1094

Fields

span: Span

TS1095

Fields

span: Span

TS1168

Fields

span: Span

Span of offending computed property.

TS1169

Fields

span: Span

Span of offending computed property.

A computed property name in an interface must refer to an expression whose type is a literal type or a ‘unique symbol’ type.

TS1183

Fields

span: Span

TS1318

Fields

span: Span

TS1319

Fields

span: Span

ExportEqualsMixedWithOtherExports

Fields

span: Span

TS2309

AnyTypeUsedAsCalleeWithTypeArgs

Fields

span: Span

InvalidLhsOfInOperator

Fields

span: Span

TS2360

InvalidRhsForInOperator

Fields

span: Span

ty: Box<Type>

WrongTypeForLhsOfNumericOperation

Fields

span: Span

TS2362

WrongTypeForRhsOfNumericOperation

Fields

span: Span

ty: Box<Type>

TS2363

TS2365

Fields

span: Span

TS2370

Fields

span: Span

WrongOverloadSignature

Fields

span: Span

TS2394

TS1166

Fields

span: Span

TS1345

Fields

span: Span

TS2353

Fields

span: Span

ConstructorImplMissingOrNotFollowedByDecl

Fields

span: Span

TS2390

FnImplMissingOrNotFollowedByDecl

Fields

span: Span

TS2391

InvalidTypeForComputedProperty

Fields

span: Span

ty: Box<Type>

TS2464

ParamPropIsNotAllowedInAmbientConstructor

Fields

span: Span

TS2369

TS2389

Fields

span: Span

TS2447

Fields

span: Span

ClassDoesNotImplementMember

Fields

span: Span

key: Box<Key>

TS2515

TS2531

Fields

span: Span

TS2567

Fields

span: Span

TS2585

Fields

span: Span

Type used as a variable, but changing target library can fix the issue.

TS2704

Fields

span: Span

InvalidLhsInInstanceOf

Fields

span: Span

ty: Box<Type>

Type of the lhs

TS2358

InvalidRhsInInstanceOf

Fields

span: Span

ty: Box<Type>

Type of the rhs

TS2359

NumericOpToSymbol

Fields

span: Span

Span of the argument.

TS2469

InvalidNumericOperand

Fields

span: Span

TS2356

UpdateOpToSymbol

Fields

span: Span

Span of the argument.

op: UpdateOp

TS2469

UnknownPropertyInObjectLiteralAssignment

Fields

span: Span

NonOverlappingTypeCast

Fields

span: Span

InvalidOperatorForLhs

Fields

span: Span

op: AssignOp

InvalidOpAssign

Fields

span: Span

op: AssignOp

lhs: Box<Type>

rhs: Box<Type>

AssignOpCannotBeApplied

Fields

span: Span

op: AssignOp

ExpectedNArgsButGotM

Fields

span: Span

min: usize

max: Option<usize>

ExpectedAtLeastNArgsButGotM

Fields

span: Span

min: usize

ExpectedAtLeastNArgsButGotMOrMore

Fields

span: Span

min: usize

ExpectedNArgsButGotMOrMore

Fields

span: Span

SpreadMustBeTupleOrPassedToRest

Fields

span: Span

TS2556

InvalidDeleteOperand

Fields

span: Span

TS2703

NoMethodNamedNext

Fields

span: Span

TS2489

NoCallablePropertyWithName

Fields

span: Span

obj: Box<Type>

key: Box<Key>

TS2489

MustBeArray

Fields

span: Span

TS2461

MustHaveSymbolIteratorThatReturnsIteratorOrMustBeArray

Fields

span: Span

TS2548

MustHaveSymbolIteratorThatReturnsIterator

Fields

span: Span

TS2488

RightHandSideMustBeObject

Fields

span: Span

ty: Box<Type>

TS2407

MustHaveSymbolAsyncIteratorThatReturnsIterator

Fields

span: Span

NoSuchConstructor

Fields

span: Span

key: Box<Key>

AbstractAndConcreteIsMixed

Fields

span: Span

TS2512

SuperCanOnlyAccessPublicAndProtectedMethod

Fields

span: Span

TS2340

ShouldBeStaticMethod

Fields

span: Span

TS2387

ShouldBeInstanceMethod

Fields

span: Span

TS2388

RestPropertyNotLast

Fields

span: Span

TS2462

ClassConstructorPrivate

Fields

span: Span

TS2673

ClassConstructorProtected

Fields

span: Span

TS2674

InvalidExtendDueToConstructorPrivate

Fields

span: Span

TS2675

DuplicatePrivateStaticInstance

Fields

span: Span

TS2804

ExportAmbientModule

Fields

span: Span

TS2668

IsTypeUnknown

Fields

span: Span

TS18046

RestTypeNotFromObject

Fields

span: Span

TS2700

UsePropBeforeInit

Fields

span: Span

obj: Option<Box<Type>>

prop: Option<Box<Key>>

TS2729

Implementations

impl ErrorKind

pub fn normalize_error_code(code: usize) -> usize

pub fn context(self, context: impl Display) -> Error

pub fn into_causes(self) -> Vec<Error>

Split error into causes.

pub fn code(&self) -> usize

TypeScript error code.

pub fn is_property_not_found(&self) -> bool

pub fn is_var_not_found(&self) -> bool

pub fn is_readonly_error(&self) -> bool

pub fn is_assign_failure(&self) -> bool

pub fn is_type_not_found(&self) -> bool

pub fn flatten(vec: Vec<Error>) -> Vec<Error>